Refactor schedule models + clean up
