+++ app/src/main/java/com/example/peterchu/watplanner/coursedetail/CourseDetailFragment.java
@@ -10,7 +10,6 @@ import android.support.v4.app.Fragment;
 import android.support.v4.view.ViewCompat;
 import android.support.v4.view.ViewPropertyAnimatorCompat;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.CardView;
 import android.support.v7.widget.Toolbar;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
@@ -22,12 +21,12 @@ import android.widget.TextView;
 import com.alamkanak.weekview.MonthLoader;
 import com.alamkanak.weekview.WeekView;
 import com.alamkanak.weekview.WeekViewEvent;
-import com.alamkanak.weekview.WeekViewLoader;
 import com.example.peterchu.watplanner.BaseView;
 import com.example.peterchu.watplanner.Calendar.WeekViewCourseEvent;
 import com.example.peterchu.watplanner.Models.Course.CourseDetails;
-import com.example.peterchu.watplanner.Models.Schedule.CourseComponent;
+import com.example.peterchu.watplanner.Models.Schedule.CourseScheduleComponent;
 import com.example.peterchu.watplanner.R;
+import com.example.peterchu.watplanner.util.ScheduleUtils;
 
 import java.util.ArrayList;
 import java.util.Calendar;
@@ -40,7 +39,7 @@ public class CourseDetailFragment extends Fragment implements BaseView<CourseDet
 
     public static final String ARG_COURSE_ID = "course_id";
 
-    private List<CourseComponent> mCourseSchedule;
+    private List<CourseScheduleComponent> mCourseSchedule;
     private CourseDetailPresenter presenter;
 
     private View rootView;
@@ -85,7 +84,7 @@ public class CourseDetailFragment extends Fragment implements BaseView<CourseDet
                 courseDetails.getAntirequisites());
     }
 
-    public void setCourseSchedule(List<CourseComponent> courseSchedule) {
+    public void setCourseSchedule(List<CourseScheduleComponent> courseSchedule) {
         mCourseSchedule = courseSchedule;
         // The week view has infinite scrolling horizontally. We have to provide the events of a
         // month every time the month changes on the week view.
@@ -93,8 +92,8 @@ public class CourseDetailFragment extends Fragment implements BaseView<CourseDet
             @Override
             public List<? extends WeekViewEvent> onMonthChange(int newYear, int newMonth) {
                 List<WeekViewEvent> events = new ArrayList<>();
-                for (CourseComponent c : mCourseSchedule) {
-                    events.addAll(c.toWeekViewEvents(newMonth));
+                for (CourseScheduleComponent c : mCourseSchedule) {
+                    events.addAll(ScheduleUtils.toWeekViewEvents(c, newMonth));
                 }
                 return events;
             }
@@ -104,10 +103,11 @@ public class CourseDetailFragment extends Fragment implements BaseView<CourseDet
         weekView.setOnEventClickListener(new WeekView.EventClickListener() {
             @Override
             public void onEventClick(WeekViewEvent event, RectF eventRect) {
-                CourseComponent courseComponent = ((WeekViewCourseEvent) event).getCourseComponent();
                 ComponentItemView componentItemView = new ComponentItemView(
-                        CourseDetailFragment.this.getContext(), courseComponent
-                );
+                        CourseDetailFragment.this.getContext(),
+                        ((WeekViewCourseEvent) event).getDay(),
+                        ((WeekViewCourseEvent) event).getCourseComponent(),
+                        ((WeekViewCourseEvent) event).getScheduledClass());
                 final Dialog d = new Dialog(CourseDetailFragment.this.getContext());
                 d.setContentView(componentItemView);
                 d.show();
